Move deprecated aliasedcorrelate function to test.
